Revert "Implement `triggers deploy` command (#5426)"
This reverts commit 9343714155d5fa71c7415457dd35ab343d047d0f.
