+++ packages/wrangler/src/__tests__/deploy.test.ts
@@ -116,7 +116,7 @@ describe("deploy", () => {
 			Worker PipelineHash:  hash9999
 			Worker Mutable PipelineID (Development ONLY!): mutableId
 			Uploaded test-name (TIMINGS)
-			Deployed test-name triggers (TIMINGS)
+			Published test-name (TIMINGS)
 			  https://test-name.test-sub-domain.workers.dev
 			Current Deployment ID: Galaxy-Class
 
@@ -157,7 +157,7 @@ describe("deploy", () => {
 			Successfully logged in.
 			Total Upload: xx KiB / gzip: xx KiB
 			Uploaded test-name (TIMINGS)
-			Deployed test-name triggers (TIMINGS)
+			Published test-name (TIMINGS)
 			  https://test-name.test-sub-domain.workers.dev
 			Current Deployment ID: Galaxy-Class
 
@@ -200,7 +200,7 @@ describe("deploy", () => {
 			Successfully logged in.
 			Total Upload: xx KiB / gzip: xx KiB
 			Uploaded test-name (TIMINGS)
-			Deployed test-name triggers (TIMINGS)
+			Published test-name (TIMINGS)
 			  https://test-name.test-sub-domain.workers.dev
 			Current Deployment ID: Galaxy-Class
 
@@ -226,7 +226,7 @@ describe("deploy", () => {
 			expect(std.out).toMatchInlineSnapshot(`
 			"Total Upload: xx KiB / gzip: xx KiB
 			Uploaded test-name (TIMINGS)
-			Deployed test-name triggers (TIMINGS)
+			Published test-name (TIMINGS)
 			  https://test-name.test-sub-domain.workers.dev
 			Current Deployment ID: Galaxy-Class
 
@@ -270,7 +270,7 @@ describe("deploy", () => {
 				expect(std.out).toMatchInlineSnapshot(`
 			"Total Upload: xx KiB / gzip: xx KiB
 			Uploaded test-name (TIMINGS)
-			Deployed test-name triggers (TIMINGS)
+			Published test-name (TIMINGS)
 			  https://test-name.test-sub-domain.workers.dev
 			Current Deployment ID: Galaxy-Class
 
@@ -298,7 +298,7 @@ describe("deploy", () => {
 				expect(std.out).toMatchInlineSnapshot(`
 			"Total Upload: xx KiB / gzip: xx KiB
 			Uploaded test-name (TIMINGS)
-			Deployed test-name triggers (TIMINGS)
+			Published test-name (TIMINGS)
 			  https://test-name.test-sub-domain.workers.dev
 			Current Deployment ID: Galaxy-Class
 
@@ -429,7 +429,7 @@ describe("deploy", () => {
 			expect(std.out).toMatchInlineSnapshot(`
 			"Total Upload: xx KiB / gzip: xx KiB
 			Uploaded test-name-some-env (TIMINGS)
-			Deployed test-name-some-env triggers (TIMINGS)
+			Published test-name-some-env (TIMINGS)
 			  https://test-name-some-env.test-sub-domain.workers.dev
 			Current Deployment ID: Galaxy-Class
 
@@ -453,7 +453,7 @@ describe("deploy", () => {
 				expect(std.out).toMatchInlineSnapshot(`
 			"Total Upload: xx KiB / gzip: xx KiB
 			Uploaded test-name (TIMINGS)
-			Deployed test-name triggers (TIMINGS)
+			Published test-name (TIMINGS)
 			  https://test-name.test-sub-domain.workers.dev
 			Current Deployment ID: Galaxy-Class
 
@@ -477,7 +477,7 @@ describe("deploy", () => {
 				expect(std.out).toMatchInlineSnapshot(`
 			"Total Upload: xx KiB / gzip: xx KiB
 			Uploaded test-name-some-env (TIMINGS)
-			Deployed test-name-some-env triggers (TIMINGS)
+			Published test-name-some-env (TIMINGS)
 			  https://test-name-some-env.test-sub-domain.workers.dev
 			Current Deployment ID: Galaxy-Class
 
@@ -501,7 +501,7 @@ describe("deploy", () => {
 				expect(std.out).toMatchInlineSnapshot(`
 			"Total Upload: xx KiB / gzip: xx KiB
 			Uploaded test-name-some-env (TIMINGS)
-			Deployed test-name-some-env triggers (TIMINGS)
+			Published test-name-some-env (TIMINGS)
 			  https://test-name-some-env.test-sub-domain.workers.dev
 			Current Deployment ID: Galaxy-Class
 
@@ -576,7 +576,7 @@ describe("deploy", () => {
 				expect(std.out).toMatchInlineSnapshot(`
 			"Total Upload: xx KiB / gzip: xx KiB
 			Uploaded test-name (TIMINGS)
-			Deployed test-name triggers (TIMINGS)
+			Published test-name (TIMINGS)
 			  https://test-name.test-sub-domain.workers.dev
 			Current Deployment ID: Galaxy-Class
 
@@ -607,7 +607,7 @@ describe("deploy", () => {
 				expect(std.out).toMatchInlineSnapshot(`
 			"Total Upload: xx KiB / gzip: xx KiB
 			Uploaded test-name (some-env) (TIMINGS)
-			Deployed test-name (some-env) triggers (TIMINGS)
+			Published test-name (some-env) (TIMINGS)
 			  https://some-env.test-name.test-sub-domain.workers.dev
 			Current Deployment ID: Galaxy-Class
 
@@ -657,7 +657,7 @@ describe("deploy", () => {
 		- Vars:
 		  - xyz: 123
 		Uploaded test-name (TIMINGS)
-		Deployed test-name triggers (TIMINGS)
+		Published test-name (TIMINGS)
 		  https://test-name.test-sub-domain.workers.dev
 		Current Deployment ID: Galaxy-Class
 
@@ -696,7 +696,7 @@ describe("deploy", () => {
 			  "info": "",
 			  "out": "Total Upload: xx KiB / gzip: xx KiB
 			Uploaded test-name (TIMINGS)
-			Deployed test-name triggers (TIMINGS)
+			Published test-name (TIMINGS)
 			  https://test-name.test-sub-domain.workers.dev
 			Current Deployment ID: Galaxy-Class
 
@@ -747,7 +747,7 @@ describe("deploy", () => {
 			  "info": "",
 			  "out": "Total Upload: xx KiB / gzip: xx KiB
 			Uploaded test-name (TIMINGS)
-			Deployed test-name triggers (TIMINGS)
+			Published test-name (TIMINGS)
 			  some-example.com/some-route/*
 			  *a-boring-website.com (zone id: 54sdf7fsda)
 			  *another-boring-website.com (zone name: some-zone.com)
@@ -793,7 +793,7 @@ describe("deploy", () => {
 			  "info": "",
 			  "out": "Total Upload: xx KiB / gzip: xx KiB
 			Uploaded test-name (TIMINGS)
-			Deployed test-name triggers (TIMINGS)
+			Published test-name (TIMINGS)
 			  partner.com/* (zone name: owned-zone.com)
 			Current Deployment ID: Galaxy-Class
 
@@ -835,7 +835,7 @@ describe("deploy", () => {
 			  "info": "",
 			  "out": "Total Upload: xx KiB / gzip: xx KiB
 			Uploaded test-name (TIMINGS)
-			Deployed test-name triggers (TIMINGS)
+			Published test-name (TIMINGS)
 			  subdomain.partner.com/* (zone name: owned-zone.com)
 			Current Deployment ID: Galaxy-Class
 
@@ -897,7 +897,7 @@ describe("deploy", () => {
 			  "info": "",
 			  "out": "Total Upload: xx KiB / gzip: xx KiB
 			Uploaded test-name (staging) (TIMINGS)
-			Deployed test-name (staging) triggers (TIMINGS)
+			Published test-name (staging) (TIMINGS)
 			  some-example.com/some-route/*
 			  *a-boring-website.com (zone id: 54sdf7fsda)
 			  *another-boring-website.com (zone name: some-zone.com)
@@ -1007,7 +1007,7 @@ describe("deploy", () => {
 			expect(std.out).toMatchInlineSnapshot(`
 			"Total Upload: xx KiB / gzip: xx KiB
 			Uploaded test-name (TIMINGS)
-			Deployed test-name triggers (TIMINGS)
+			Published test-name (TIMINGS)
 			  example.com/some-route/*
 			Current Deployment ID: Galaxy-Class
 
@@ -1040,9 +1040,9 @@ describe("deploy", () => {
 			});
 			await expect(runWrangler("deploy ./index --env=staging")).rejects
 				.toThrowErrorMatchingInlineSnapshot(`
-			"Service environments combined with an API token that doesn't have 'All Zones' permissions is not supported.
-			Either turn off service environments by setting \`legacy_env = true\`, creating an API token with 'All Zones' permissions, or logging in via OAuth"
-		`);
+			              "Service environments combined with an API token that doesn't have 'All Zones' permissions is not supported.
+			              Either turn off service environments by setting \`legacy_env = true\`, creating an API token with 'All Zones' permissions, or logging in via OAuth"
+		            `);
 		});
 
 		describe("custom domains", () => {
@@ -1293,7 +1293,7 @@ Update them to point to this script instead?`,
 			expect(std.out).toMatchInlineSnapshot(`
 			"Total Upload: xx KiB / gzip: xx KiB
 			Uploaded test-name (TIMINGS)
-			Deployed test-name triggers (TIMINGS)
+			Published test-name (TIMINGS)
 			  https://test-name.test-sub-domain.workers.dev
 			Current Deployment ID: Galaxy-Class
 
@@ -1314,7 +1314,7 @@ Update them to point to this script instead?`,
 			expect(std.out).toMatchInlineSnapshot(`
 			"Total Upload: xx KiB / gzip: xx KiB
 			Uploaded test-name (TIMINGS)
-			Deployed test-name triggers (TIMINGS)
+			Published test-name (TIMINGS)
 			  https://test-name.test-sub-domain.workers.dev
 			Current Deployment ID: Galaxy-Class
 
@@ -1335,7 +1335,7 @@ Update them to point to this script instead?`,
 			expect(std.out).toMatchInlineSnapshot(`
 			"Total Upload: xx KiB / gzip: xx KiB
 			Uploaded test-name (TIMINGS)
-			Deployed test-name triggers (TIMINGS)
+			Published test-name (TIMINGS)
 			  https://test-name.test-sub-domain.workers.dev
 			Current Deployment ID: Galaxy-Class
 
@@ -1358,7 +1358,7 @@ Update them to point to this script instead?`,
 			expect(std.out).toMatchInlineSnapshot(`
 			"Total Upload: xx KiB / gzip: xx KiB
 			Uploaded test-name (TIMINGS)
-			Deployed test-name triggers (TIMINGS)
+			Published test-name (TIMINGS)
 			  https://test-name.test-sub-domain.workers.dev
 			Current Deployment ID: Galaxy-Class
 
@@ -1379,7 +1379,7 @@ Update them to point to this script instead?`,
 			expect(std.out).toMatchInlineSnapshot(`
 			"Total Upload: xx KiB / gzip: xx KiB
 			Uploaded test-name (TIMINGS)
-			Deployed test-name triggers (TIMINGS)
+			Published test-name (TIMINGS)
 			  https://test-name.test-sub-domain.workers.dev
 			Current Deployment ID: Galaxy-Class
 
@@ -1419,7 +1419,7 @@ Update them to point to this script instead?`,
 			expect(std.out).toMatchInlineSnapshot(`
 			"Total Upload: xx KiB / gzip: xx KiB
 			Uploaded test-name (TIMINGS)
-			Deployed test-name triggers (TIMINGS)
+			Published test-name (TIMINGS)
 			  https://test-name.test-sub-domain.workers.dev
 			Current Deployment ID: Galaxy-Class
 
@@ -1473,7 +1473,7 @@ Update them to point to this script instead?`,
 			expect(std.out).toMatchInlineSnapshot(`
 			"Total Upload: xx KiB / gzip: xx KiB
 			Uploaded test-name (TIMINGS)
-			Deployed test-name triggers (TIMINGS)
+			Published test-name (TIMINGS)
 			  https://test-name.test-sub-domain.workers.dev
 			Current Deployment ID: Galaxy-Class
 
@@ -1496,7 +1496,7 @@ Update them to point to this script instead?`,
 			expect(std.out).toMatchInlineSnapshot(`
 			"Total Upload: xx KiB / gzip: xx KiB
 			Uploaded test-name (TIMINGS)
-			Deployed test-name triggers (TIMINGS)
+			Published test-name (TIMINGS)
 			  https://test-name.test-sub-domain.workers.dev
 			Current Deployment ID: Galaxy-Class
 
@@ -1531,7 +1531,7 @@ export default{
 			expect(std.out).toMatchInlineSnapshot(`
 			"Total Upload: xx KiB / gzip: xx KiB
 			Uploaded test-name (TIMINGS)
-			Deployed test-name triggers (TIMINGS)
+			Published test-name (TIMINGS)
 			  https://test-name.test-sub-domain.workers.dev
 			Current Deployment ID: Galaxy-Class
 
@@ -1560,7 +1560,7 @@ export default{
 			expect(std.out).toMatchInlineSnapshot(`
 			"Total Upload: xx KiB / gzip: xx KiB
 			Uploaded test-name (TIMINGS)
-			Deployed test-name triggers (TIMINGS)
+			Published test-name (TIMINGS)
 			  https://test-name.test-sub-domain.workers.dev
 			Current Deployment ID: Galaxy-Class
 
@@ -1581,7 +1581,7 @@ export default{
 			expect(std.out).toMatchInlineSnapshot(`
 			"Total Upload: xx KiB / gzip: xx KiB
 			Uploaded test-name (TIMINGS)
-			Deployed test-name triggers (TIMINGS)
+			Published test-name (TIMINGS)
 			  https://test-name.test-sub-domain.workers.dev
 			Current Deployment ID: Galaxy-Class
 
@@ -1681,7 +1681,7 @@ addEventListener('fetch', event => {});`
 			expect(std.out).toMatchInlineSnapshot(`
 			"Total Upload: xx KiB / gzip: xx KiB
 			Uploaded test-name (TIMINGS)
-			Deployed test-name triggers (TIMINGS)
+			Published test-name (TIMINGS)
 			  https://test-name.test-sub-domain.workers.dev
 			Current Deployment ID: Galaxy-Class
 
@@ -1766,7 +1766,7 @@ addEventListener('fetch', event => {});`
 			  "out": "↗️  Done syncing assets
 			Total Upload: xx KiB / gzip: xx KiB
 			Uploaded test-name (TIMINGS)
-			Deployed test-name triggers (TIMINGS)
+			Published test-name (TIMINGS)
 			  https://test-name.test-sub-domain.workers.dev
 			Current Deployment ID: Galaxy-Class
 
@@ -1826,7 +1826,7 @@ addEventListener('fetch', event => {});`
 			"↗️  Done syncing assets
 			Total Upload: xx KiB / gzip: xx KiB
 			Uploaded test-name (TIMINGS)
-			Deployed test-name triggers (TIMINGS)
+			Published test-name (TIMINGS)
 			  https://test-name.test-sub-domain.workers.dev
 			Current Deployment ID: Galaxy-Class
 
@@ -1918,7 +1918,7 @@ addEventListener('fetch', event => {});`
 			  "out": "↗️  Done syncing assets
 			Total Upload: xx KiB / gzip: xx KiB
 			Uploaded test-name (TIMINGS)
-			Deployed test-name triggers (TIMINGS)
+			Published test-name (TIMINGS)
 			  https://test-name.test-sub-domain.workers.dev
 			Current Deployment ID: Galaxy-Class
 
@@ -2087,7 +2087,7 @@ addEventListener('fetch', event => {});`
 			"↗️  Done syncing assets
 			Total Upload: xx KiB / gzip: xx KiB
 			Uploaded test-name (TIMINGS)
-			Deployed test-name triggers (TIMINGS)
+			Published test-name (TIMINGS)
 			  https://test-name.test-sub-domain.workers.dev
 			Current Deployment ID: Galaxy-Class
 
@@ -2133,7 +2133,7 @@ addEventListener('fetch', event => {});`
 			  "out": "↗️  Done syncing assets
 			Total Upload: xx KiB / gzip: xx KiB
 			Uploaded test-name (TIMINGS)
-			Deployed test-name triggers (TIMINGS)
+			Published test-name (TIMINGS)
 			  https://test-name.test-sub-domain.workers.dev
 			Current Deployment ID: Galaxy-Class
 
@@ -2323,7 +2323,7 @@ addEventListener('fetch', event => {});`
 			  "out": "↗️  Done syncing assets
 			Total Upload: xx KiB / gzip: xx KiB
 			Uploaded test-name (TIMINGS)
-			Deployed test-name triggers (TIMINGS)
+			Published test-name (TIMINGS)
 			  https://test-name.test-sub-domain.workers.dev
 			Current Deployment ID: Galaxy-Class
 
@@ -2375,7 +2375,7 @@ addEventListener('fetch', event => {});`
 			  "out": "↗️  Done syncing assets
 			Total Upload: xx KiB / gzip: xx KiB
 			Uploaded test-name (TIMINGS)
-			Deployed test-name triggers (TIMINGS)
+			Published test-name (TIMINGS)
 			  https://test-name.test-sub-domain.workers.dev
 			Current Deployment ID: Galaxy-Class
 
@@ -2439,7 +2439,7 @@ addEventListener('fetch', event => {});`
 			"↗️  Done syncing assets
 			Total Upload: xx KiB / gzip: xx KiB
 			Uploaded test-name (TIMINGS)
-			Deployed test-name triggers (TIMINGS)
+			Published test-name (TIMINGS)
 			  https://test-name.test-sub-domain.workers.dev
 			Current Deployment ID: Galaxy-Class
 
@@ -2504,7 +2504,7 @@ addEventListener('fetch', event => {});`
 			"↗️  Done syncing assets
 			Total Upload: xx KiB / gzip: xx KiB
 			Uploaded test-name (TIMINGS)
-			Deployed test-name triggers (TIMINGS)
+			Published test-name (TIMINGS)
 			  https://test-name.test-sub-domain.workers.dev
 			Current Deployment ID: Galaxy-Class
 
@@ -2602,7 +2602,7 @@ addEventListener('fetch', event => {});`
 			↗️  Done syncing assets
 			Total Upload: xx KiB / gzip: xx KiB
 			Uploaded test-name (TIMINGS)
-			Deployed test-name triggers (TIMINGS)
+			Published test-name (TIMINGS)
 			  https://test-name.test-sub-domain.workers.dev
 			Current Deployment ID: Galaxy-Class
 
@@ -2659,7 +2659,7 @@ addEventListener('fetch', event => {});`
 			"↗️  Done syncing assets
 			Total Upload: xx KiB / gzip: xx KiB
 			Uploaded test-name (some-env) (TIMINGS)
-			Deployed test-name (some-env) triggers (TIMINGS)
+			Published test-name (some-env) (TIMINGS)
 			  https://some-env.test-name.test-sub-domain.workers.dev
 			Current Deployment ID: Galaxy-Class
 
@@ -2717,7 +2717,7 @@ addEventListener('fetch', event => {});`
 			"↗️  Done syncing assets
 			Total Upload: xx KiB / gzip: xx KiB
 			Uploaded test-name-some-env (TIMINGS)
-			Deployed test-name-some-env triggers (TIMINGS)
+			Published test-name-some-env (TIMINGS)
 			  https://test-name-some-env.test-sub-domain.workers.dev
 			Current Deployment ID: Galaxy-Class
 
@@ -2760,7 +2760,7 @@ addEventListener('fetch', event => {});`
 			"↗️  Done syncing assets
 			Total Upload: xx KiB / gzip: xx KiB
 			Uploaded test-name (TIMINGS)
-			Deployed test-name triggers (TIMINGS)
+			Published test-name (TIMINGS)
 			  https://test-name.test-sub-domain.workers.dev
 			Current Deployment ID: Galaxy-Class
 
@@ -2809,7 +2809,7 @@ addEventListener('fetch', event => {});`
 			"↗️  Done syncing assets
 			Total Upload: xx KiB / gzip: xx KiB
 			Uploaded test-name (TIMINGS)
-			Deployed test-name triggers (TIMINGS)
+			Published test-name (TIMINGS)
 			  https://test-name.test-sub-domain.workers.dev
 			Current Deployment ID: Galaxy-Class
 
@@ -2858,7 +2858,7 @@ addEventListener('fetch', event => {});`
 			"↗️  Done syncing assets
 			Total Upload: xx KiB / gzip: xx KiB
 			Uploaded test-name (TIMINGS)
-			Deployed test-name triggers (TIMINGS)
+			Published test-name (TIMINGS)
 			  https://test-name.test-sub-domain.workers.dev
 			Current Deployment ID: Galaxy-Class
 
@@ -2908,7 +2908,7 @@ addEventListener('fetch', event => {});`
 			"↗️  Done syncing assets
 			Total Upload: xx KiB / gzip: xx KiB
 			Uploaded test-name (TIMINGS)
-			Deployed test-name triggers (TIMINGS)
+			Published test-name (TIMINGS)
 			  https://test-name.test-sub-domain.workers.dev
 			Current Deployment ID: Galaxy-Class
 
@@ -2958,7 +2958,7 @@ addEventListener('fetch', event => {});`
 			"↗️  Done syncing assets
 			Total Upload: xx KiB / gzip: xx KiB
 			Uploaded test-name (TIMINGS)
-			Deployed test-name triggers (TIMINGS)
+			Published test-name (TIMINGS)
 			  https://test-name.test-sub-domain.workers.dev
 			Current Deployment ID: Galaxy-Class
 
@@ -3008,7 +3008,7 @@ addEventListener('fetch', event => {});`
 			"↗️  Done syncing assets
 			Total Upload: xx KiB / gzip: xx KiB
 			Uploaded test-name (TIMINGS)
-			Deployed test-name triggers (TIMINGS)
+			Published test-name (TIMINGS)
 			  https://test-name.test-sub-domain.workers.dev
 			Current Deployment ID: Galaxy-Class
 
@@ -3058,7 +3058,7 @@ addEventListener('fetch', event => {});`
 			"↗️  Done syncing assets
 			Total Upload: xx KiB / gzip: xx KiB
 			Uploaded test-name (TIMINGS)
-			Deployed test-name triggers (TIMINGS)
+			Published test-name (TIMINGS)
 			  https://test-name.test-sub-domain.workers.dev
 			Current Deployment ID: Galaxy-Class
 
@@ -3110,7 +3110,7 @@ addEventListener('fetch', event => {});`
 			"↗️  Done syncing assets
 			Total Upload: xx KiB / gzip: xx KiB
 			Uploaded test-name (TIMINGS)
-			Deployed test-name triggers (TIMINGS)
+			Published test-name (TIMINGS)
 			  https://test-name.test-sub-domain.workers.dev
 			Current Deployment ID: Galaxy-Class
 
@@ -3166,7 +3166,7 @@ addEventListener('fetch', event => {});`
 			"↗️  Done syncing assets
 			Total Upload: xx KiB / gzip: xx KiB
 			Uploaded test-name (TIMINGS)
-			Deployed test-name triggers (TIMINGS)
+			Published test-name (TIMINGS)
 			  https://test-name.test-sub-domain.workers.dev
 			Current Deployment ID: Galaxy-Class
 
@@ -3278,7 +3278,7 @@ addEventListener('fetch', event => {});`
 			"↗️  Done syncing assets
 			Total Upload: xx KiB / gzip: xx KiB
 			Uploaded test-name (TIMINGS)
-			Deployed test-name triggers (TIMINGS)
+			Published test-name (TIMINGS)
 			  https://test-name.test-sub-domain.workers.dev
 			Current Deployment ID: Galaxy-Class
 
@@ -3419,7 +3419,7 @@ addEventListener('fetch', event => {});`
 			"↗️  Done syncing assets
 			Total Upload: xx KiB / gzip: xx KiB
 			Uploaded test-name (TIMINGS)
-			Deployed test-name triggers (TIMINGS)
+			Published test-name (TIMINGS)
 			  https://test-name.test-sub-domain.workers.dev
 			Current Deployment ID: Galaxy-Class
 
@@ -3481,7 +3481,7 @@ addEventListener('fetch', event => {});`
 			"↗️  Done syncing assets
 			Total Upload: xx KiB / gzip: xx KiB
 			Uploaded test-name (TIMINGS)
-			Deployed test-name triggers (TIMINGS)
+			Published test-name (TIMINGS)
 			  https://test-name.test-sub-domain.workers.dev
 			Current Deployment ID: Galaxy-Class
 
@@ -3528,7 +3528,7 @@ addEventListener('fetch', event => {});`
 			  "out": "↗️  Done syncing assets
 			Total Upload: xx KiB / gzip: xx KiB
 			Uploaded test-name (TIMINGS)
-			Deployed test-name triggers (TIMINGS)
+			Published test-name (TIMINGS)
 			  https://test-name.test-sub-domain.workers.dev
 			Current Deployment ID: Galaxy-Class
 
@@ -3576,7 +3576,7 @@ addEventListener('fetch', event => {});`
 			  "out": "↗️  Done syncing assets
 			Total Upload: xx KiB / gzip: xx KiB
 			Uploaded test-name (TIMINGS)
-			Deployed test-name triggers (TIMINGS)
+			Published test-name (TIMINGS)
 			  https://test-name.test-sub-domain.workers.dev
 			Current Deployment ID: Galaxy-Class
 
@@ -3797,7 +3797,7 @@ addEventListener('fetch', event => {});`
 			  "out": "↗️  Done syncing assets
 			Total Upload: xx KiB / gzip: xx KiB
 			Uploaded test-name (TIMINGS)
-			Deployed test-name triggers (TIMINGS)
+			Published test-name (TIMINGS)
 			  https://test-name.test-sub-domain.workers.dev
 			Current Deployment ID: Galaxy-Class
 
@@ -3951,7 +3951,7 @@ addEventListener('fetch', event => {});`
 			expect(std.out).toMatchInlineSnapshot(`
 			"Total Upload: xx KiB / gzip: xx KiB
 			Uploaded test-name (TIMINGS)
-			Deployed test-name triggers (TIMINGS)
+			Published test-name (TIMINGS)
 			  https://test-name.test-sub-domain.workers.dev
 			Current Deployment ID: Galaxy-Class
 
@@ -3975,7 +3975,7 @@ addEventListener('fetch', event => {});`
 			expect(std.out).toMatchInlineSnapshot(`
 			"Total Upload: xx KiB / gzip: xx KiB
 			Uploaded test-name (TIMINGS)
-			Deployed test-name triggers (TIMINGS)
+			Published test-name (TIMINGS)
 			  https://test-name.test-sub-domain.workers.dev
 			Current Deployment ID: Galaxy-Class
 
@@ -3998,7 +3998,7 @@ addEventListener('fetch', event => {});`
 			expect(std.out).toMatchInlineSnapshot(`
 			"Total Upload: xx KiB / gzip: xx KiB
 			Uploaded test-name (TIMINGS)
-			Deployed test-name triggers (TIMINGS)
+			Published test-name (TIMINGS)
 			  https://test-name.test-sub-domain.workers.dev
 			Current Deployment ID: Galaxy-Class
 
@@ -4131,7 +4131,7 @@ addEventListener('fetch', event => {});`
 			expect(std.out).toMatchInlineSnapshot(`
 			"Total Upload: xx KiB / gzip: xx KiB
 			Uploaded test-name (dev) (TIMINGS)
-			Deployed test-name (dev) triggers (TIMINGS)
+			Published test-name (dev) (TIMINGS)
 			  https://dev.test-name.test-sub-domain.workers.dev
 			Current Deployment ID: Galaxy-Class
 
@@ -4162,7 +4162,7 @@ addEventListener('fetch', event => {});`
 			expect(std.out).toMatchInlineSnapshot(`
 			"Total Upload: xx KiB / gzip: xx KiB
 			Uploaded test-name (dev) (TIMINGS)
-			Deployed test-name (dev) triggers (TIMINGS)
+			Published test-name (dev) (TIMINGS)
 			  https://dev.test-name.test-sub-domain.workers.dev
 			Current Deployment ID: Galaxy-Class
 
@@ -4194,7 +4194,7 @@ addEventListener('fetch', event => {});`
 			expect(std.out).toMatchInlineSnapshot(`
 			"Total Upload: xx KiB / gzip: xx KiB
 			Uploaded test-name (dev) (TIMINGS)
-			Deployed test-name (dev) triggers (TIMINGS)
+			Published test-name (dev) (TIMINGS)
 			  https://dev.test-name.test-sub-domain.workers.dev
 			Current Deployment ID: Galaxy-Class
 
@@ -4229,7 +4229,7 @@ addEventListener('fetch', event => {});`
 			expect(std.out).toMatchInlineSnapshot(`
 			"Total Upload: xx KiB / gzip: xx KiB
 			Uploaded test-name (dev) (TIMINGS)
-			Deployed test-name (dev) triggers (TIMINGS)
+			Published test-name (dev) (TIMINGS)
 			  https://dev.test-name.test-sub-domain.workers.dev
 			Current Deployment ID: Galaxy-Class
 
@@ -4266,7 +4266,7 @@ addEventListener('fetch', event => {});`
 			expect(std.out).toMatchInlineSnapshot(`
 			"Total Upload: xx KiB / gzip: xx KiB
 			Uploaded test-name (dev) (TIMINGS)
-			Deployed test-name (dev) triggers (TIMINGS)
+			Published test-name (dev) (TIMINGS)
 			  https://dev.test-name.test-sub-domain.workers.dev
 			Current Deployment ID: Galaxy-Class
 
@@ -4330,7 +4330,7 @@ addEventListener('fetch', event => {});`
 			expect(std.out).toMatchInlineSnapshot(`
 			"Total Upload: xx KiB / gzip: xx KiB
 			Uploaded test-name (TIMINGS)
-			Deployed test-name triggers (TIMINGS)
+			Published test-name (TIMINGS)
 			  https://test-name.test-sub-domain.workers.dev
 			Current Deployment ID: Galaxy-Class
 
@@ -4352,7 +4352,7 @@ addEventListener('fetch', event => {});`
 			expect(std.out).toMatchInlineSnapshot(`
 			"Total Upload: xx KiB / gzip: xx KiB
 			Uploaded test-name (TIMINGS)
-			Deployed test-name triggers (TIMINGS)
+			Published test-name (TIMINGS)
 			  https://test-name.test-sub-domain.workers.dev
 			Current Deployment ID: Galaxy-Class
 
@@ -4432,7 +4432,7 @@ addEventListener('fetch', event => {});`
 			expect(std.out).toMatchInlineSnapshot(`
 			"Total Upload: xx KiB / gzip: xx KiB
 			Uploaded test-name (TIMINGS)
-			Deployed test-name triggers (TIMINGS)
+			Published test-name (TIMINGS)
 			  http://example.com/*
 			Current Deployment ID: Galaxy-Class
 
@@ -4470,7 +4470,7 @@ addEventListener('fetch', event => {});`
 			expect(std.out).toMatchInlineSnapshot(`
 			"Total Upload: xx KiB / gzip: xx KiB
 			Uploaded test-name-production (TIMINGS)
-			Deployed test-name-production triggers (TIMINGS)
+			Published test-name-production (TIMINGS)
 			  http://production.example.com/*
 			Current Deployment ID: Galaxy-Class
 
@@ -4507,7 +4507,7 @@ addEventListener('fetch', event => {});`
 			expect(std.out).toMatchInlineSnapshot(`
 			"Total Upload: xx KiB / gzip: xx KiB
 			Uploaded test-name-production (TIMINGS)
-			Deployed test-name-production triggers (TIMINGS)
+			Published test-name-production (TIMINGS)
 			  http://production.example.com/*
 			Current Deployment ID: Galaxy-Class
 
@@ -4537,7 +4537,7 @@ addEventListener('fetch', event => {});`
 			expect(std.out).toMatchInlineSnapshot(`
 			"Total Upload: xx KiB / gzip: xx KiB
 			Uploaded test-name (TIMINGS)
-			Deployed test-name triggers (TIMINGS)
+			Published test-name (TIMINGS)
 			  https://test-name.test-sub-domain.workers.dev
 			  http://example.com/*
 			Current Deployment ID: Galaxy-Class
@@ -4576,7 +4576,7 @@ addEventListener('fetch', event => {});`
 			expect(std.out).toMatchInlineSnapshot(`
 			"Total Upload: xx KiB / gzip: xx KiB
 			Uploaded test-name-production (TIMINGS)
-			Deployed test-name-production triggers (TIMINGS)
+			Published test-name-production (TIMINGS)
 			  https://test-name-production.test-sub-domain.workers.dev
 			  http://production.example.com/*
 			Current Deployment ID: Galaxy-Class
@@ -4615,7 +4615,7 @@ addEventListener('fetch', event => {});`
 			expect(std.out).toMatchInlineSnapshot(`
 			"Total Upload: xx KiB / gzip: xx KiB
 			Uploaded test-name-production (TIMINGS)
-			Deployed test-name-production triggers (TIMINGS)
+			Published test-name-production (TIMINGS)
 			  https://test-name-production.test-sub-domain.workers.dev
 			  http://production.example.com/*
 			Current Deployment ID: Galaxy-Class
@@ -4654,7 +4654,7 @@ addEventListener('fetch', event => {});`
 			expect(std.out).toMatchInlineSnapshot(`
 			"Total Upload: xx KiB / gzip: xx KiB
 			Uploaded test-name-production (TIMINGS)
-			Deployed test-name-production triggers (TIMINGS)
+			Published test-name-production (TIMINGS)
 			  http://production.example.com/*
 			Current Deployment ID: Galaxy-Class
 
@@ -4692,7 +4692,7 @@ addEventListener('fetch', event => {});`
 			expect(std.out).toMatchInlineSnapshot(`
 			"Total Upload: xx KiB / gzip: xx KiB
 			Uploaded test-name-production (TIMINGS)
-			Deployed test-name-production triggers (TIMINGS)
+			Published test-name-production (TIMINGS)
 			  http://production.example.com/*
 			Current Deployment ID: Galaxy-Class
 
@@ -4847,7 +4847,7 @@ addEventListener('fetch', event => {});`
 			"Running custom build: node -e \\"4+4; require('fs').writeFileSync('index.js', 'export default { fetch(){ return new Response(123) } }')\\"
 			Total Upload: xx KiB / gzip: xx KiB
 			Uploaded test-name (TIMINGS)
-			Deployed test-name triggers (TIMINGS)
+			Published test-name (TIMINGS)
 			  https://test-name.test-sub-domain.workers.dev
 			Current Deployment ID: Galaxy-Class
 
@@ -4876,7 +4876,7 @@ addEventListener('fetch', event => {});`
 			"Running custom build: echo \\"export default { fetch(){ return new Response(123) } }\\" > index.js
 			Total Upload: xx KiB / gzip: xx KiB
 			Uploaded test-name (TIMINGS)
-			Deployed test-name triggers (TIMINGS)
+			Published test-name (TIMINGS)
 			  https://test-name.test-sub-domain.workers.dev
 			Current Deployment ID: Galaxy-Class
 
@@ -4984,7 +4984,7 @@ addEventListener('fetch', event => {});`
 			expect(std.out).toMatchInlineSnapshot(`
 			"Total Upload: xx KiB / gzip: xx KiB
 			Uploaded test-name (TIMINGS)
-			Deployed test-name triggers (TIMINGS)
+			Published test-name (TIMINGS)
 			  https://test-name.test-sub-domain.workers.dev
 			Current Deployment ID: Galaxy-Class
 
@@ -5027,7 +5027,7 @@ addEventListener('fetch', event => {});`
 			expect(std.out).toMatchInlineSnapshot(`
 			"Total Upload: xx KiB / gzip: xx KiB
 			Uploaded test-name (testEnv) (TIMINGS)
-			Deployed test-name (testEnv) triggers (TIMINGS)
+			Published test-name (testEnv) (TIMINGS)
 			  https://testEnv.test-name.test-sub-domain.workers.dev
 			Current Deployment ID: Galaxy-Class
 
@@ -5058,7 +5058,7 @@ addEventListener('fetch', event => {});`
 			- Durable Objects:
 			  - SOMENAME: SomeClass
 			Uploaded test-name (TIMINGS)
-			Deployed test-name triggers (TIMINGS)
+			Published test-name (TIMINGS)
 			  https://test-name.test-sub-domain.workers.dev
 			Current Deployment ID: Galaxy-Class
 
@@ -5112,7 +5112,7 @@ addEventListener('fetch', event => {});`
 			- Durable Objects:
 			  - SOMENAME: SomeClass (defined in some-script)
 			Uploaded test-name (TIMINGS)
-			Deployed test-name triggers (TIMINGS)
+			Published test-name (TIMINGS)
 			  https://test-name.test-sub-domain.workers.dev
 			Current Deployment ID: Galaxy-Class
 
@@ -5160,7 +5160,7 @@ addEventListener('fetch', event => {});`
 			  - SOMENAME: SomeClass
 			  - SOMEOTHERNAME: SomeOtherClass
 			Uploaded test-name (TIMINGS)
-			Deployed test-name triggers (TIMINGS)
+			Published test-name (TIMINGS)
 			  https://test-name.test-sub-domain.workers.dev
 			Current Deployment ID: Galaxy-Class
 
@@ -5216,7 +5216,7 @@ addEventListener('fetch', event => {});`
 			  - SOMENAME: SomeClass
 			  - SOMEOTHERNAME: SomeOtherClass
 			Uploaded test-name (TIMINGS)
-			Deployed test-name triggers (TIMINGS)
+			Published test-name (TIMINGS)
 			  https://test-name.test-sub-domain.workers.dev
 			Current Deployment ID: Galaxy-Class
 
@@ -5265,7 +5265,7 @@ addEventListener('fetch', event => {});`
 			  - SOMENAME: SomeClass
 			  - SOMEOTHERNAME: SomeOtherClass
 			Uploaded test-name (TIMINGS)
-			Deployed test-name triggers (TIMINGS)
+			Published test-name (TIMINGS)
 			  https://test-name.test-sub-domain.workers.dev
 			Current Deployment ID: Galaxy-Class
 
@@ -5315,7 +5315,7 @@ addEventListener('fetch', event => {});`
 			  - SOMENAME: SomeClass
 			  - SOMEOTHERNAME: SomeOtherClass
 			Uploaded test-name (TIMINGS)
-			Deployed test-name triggers (TIMINGS)
+			Published test-name (TIMINGS)
 			  https://test-name.test-sub-domain.workers.dev
 			Current Deployment ID: Galaxy-Class
 
@@ -5382,7 +5382,7 @@ addEventListener('fetch', event => {});`
 			  - SOMENAME: SomeClass
 			  - SOMEOTHERNAME: SomeOtherClass
 			Uploaded test-name (xyz) (TIMINGS)
-			Deployed test-name (xyz) triggers (TIMINGS)
+			Published test-name (xyz) (TIMINGS)
 			  https://xyz.test-name.test-sub-domain.workers.dev
 			Current Deployment ID: Galaxy-Class
 
@@ -5448,7 +5448,7 @@ addEventListener('fetch', event => {});`
 			  - SOMENAME: SomeClass
 			  - SOMEOTHERNAME: SomeOtherClass
 			Uploaded test-name (TIMINGS)
-			Deployed test-name triggers (TIMINGS)
+			Published test-name (TIMINGS)
 			  https://test-name.test-sub-domain.workers.dev
 			Current Deployment ID: Galaxy-Class
 
@@ -5522,7 +5522,7 @@ addEventListener('fetch', event => {});`
 			  - SOMENAME: SomeClass
 			  - SOMEOTHERNAME: SomeOtherClass
 			Uploaded test-name (xyz) (TIMINGS)
-			Deployed test-name (xyz) triggers (TIMINGS)
+			Published test-name (xyz) (TIMINGS)
 			  https://xyz.test-name.test-sub-domain.workers.dev
 			Current Deployment ID: Galaxy-Class
 
@@ -5561,7 +5561,7 @@ addEventListener('fetch', event => {});`
 			expect(std.out).toMatchInlineSnapshot(`
 			"Total Upload: xx KiB / gzip: xx KiB
 			Uploaded test-name (TIMINGS)
-			Deployed test-name triggers (TIMINGS)
+			Published test-name (TIMINGS)
 			  https://test-name.test-sub-domain.workers.dev
 			Current Deployment ID: Galaxy-Class
 
@@ -5587,7 +5587,7 @@ addEventListener('fetch', event => {});`
 			expect(std.out).toMatchInlineSnapshot(`
 			"Total Upload: xx KiB / gzip: xx KiB
 			Uploaded test-name (TIMINGS)
-			Deployed test-name triggers (TIMINGS)
+			Published test-name (TIMINGS)
 			  https://test-name.test-sub-domain.workers.dev
 			Current Deployment ID: Galaxy-Class
 
@@ -5850,7 +5850,7 @@ addEventListener('fetch', event => {});`
 			  - extra_data: \\"interesting value\\"
 			  - more_data: \\"dubious value\\"
 			Uploaded test-name (TIMINGS)
-			Deployed test-name triggers (TIMINGS)
+			Published test-name (TIMINGS)
 			  https://test-name.test-sub-domain.workers.dev
 			Current Deployment ID: Galaxy-Class
 
@@ -6268,7 +6268,7 @@ addEventListener('fetch', event => {});`
 			- Wasm Modules:
 			  - TESTWASMNAME: path/to/test.wasm
 			Uploaded test-name (TIMINGS)
-			Deployed test-name triggers (TIMINGS)
+			Published test-name (TIMINGS)
 			  https://test-name.test-sub-domain.workers.dev
 			Current Deployment ID: Galaxy-Class
 
@@ -6339,7 +6339,7 @@ addEventListener('fetch', event => {});`
 			- Wasm Modules:
 			  - TESTWASMNAME: path/to/and/the/path/to/test.wasm
 			Uploaded test-name (TIMINGS)
-			Deployed test-name triggers (TIMINGS)
+			Published test-name (TIMINGS)
 			  https://test-name.test-sub-domain.workers.dev
 			Current Deployment ID: Galaxy-Class
 
@@ -6376,7 +6376,7 @@ addEventListener('fetch', event => {});`
 				expect(std.out).toMatchInlineSnapshot(`
 			"Total Upload: xx KiB / gzip: xx KiB
 			Uploaded test-name (TIMINGS)
-			Deployed test-name triggers (TIMINGS)
+			Published test-name (TIMINGS)
 			  https://test-name.test-sub-domain.workers.dev
 			Current Deployment ID: Galaxy-Class
 
@@ -6417,7 +6417,7 @@ addEventListener('fetch', event => {});`
 			- Text Blobs:
 			  - TESTTEXTBLOBNAME: path/to/text.file
 			Uploaded test-name (TIMINGS)
-			Deployed test-name triggers (TIMINGS)
+			Published test-name (TIMINGS)
 			  https://test-name.test-sub-domain.workers.dev
 			Current Deployment ID: Galaxy-Class
 
@@ -6492,7 +6492,7 @@ addEventListener('fetch', event => {});`
 			- Text Blobs:
 			  - TESTTEXTBLOBNAME: path/to/and/the/path/to/text.file
 			Uploaded test-name (TIMINGS)
-			Deployed test-name triggers (TIMINGS)
+			Published test-name (TIMINGS)
 			  https://test-name.test-sub-domain.workers.dev
 			Current Deployment ID: Galaxy-Class
 
@@ -6533,7 +6533,7 @@ addEventListener('fetch', event => {});`
 			- Data Blobs:
 			  - TESTDATABLOBNAME: path/to/data.bin
 			Uploaded test-name (TIMINGS)
-			Deployed test-name triggers (TIMINGS)
+			Published test-name (TIMINGS)
 			  https://test-name.test-sub-domain.workers.dev
 			Current Deployment ID: Galaxy-Class
 
@@ -6608,7 +6608,7 @@ addEventListener('fetch', event => {});`
 			- Data Blobs:
 			  - TESTDATABLOBNAME: path/to/and/the/path/to/data.bin
 			Uploaded test-name (TIMINGS)
-			Deployed test-name triggers (TIMINGS)
+			Published test-name (TIMINGS)
 			  https://test-name.test-sub-domain.workers.dev
 			Current Deployment ID: Galaxy-Class
 
@@ -6655,7 +6655,7 @@ addEventListener('fetch', event => {});`
 			 \\"id\\": 123
 			}
 			Uploaded test-name (TIMINGS)
-			Deployed test-name triggers (TIMINGS)
+			Published test-name (TIMINGS)
 			  https://test-name.test-sub-domain.workers.dev
 			Current Deployment ID: Galaxy-Class
 
@@ -6683,7 +6683,7 @@ addEventListener('fetch', event => {});`
 			  - TEXT: \\"(hidden)\\"
 			  - COUNT: \\"(hidden)\\"
 			Uploaded test-name (TIMINGS)
-			Deployed test-name triggers (TIMINGS)
+			Published test-name (TIMINGS)
 			  https://test-name.test-sub-domain.workers.dev
 			Current Deployment ID: Galaxy-Class
 
@@ -6715,7 +6715,7 @@ addEventListener('fetch', event => {});`
 			- R2 Buckets:
 			  - FOO: foo-bucket
 			Uploaded test-name (TIMINGS)
-			Deployed test-name triggers (TIMINGS)
+			Published test-name (TIMINGS)
 			  https://test-name.test-sub-domain.workers.dev
 			Current Deployment ID: Galaxy-Class
 
@@ -6768,7 +6768,7 @@ addEventListener('fetch', event => {});`
 			  - httplogs: httplogs
 			  - trace: trace
 			Uploaded test-name (TIMINGS)
-			Deployed test-name triggers (TIMINGS)
+			Published test-name (TIMINGS)
 			  https://test-name.test-sub-domain.workers.dev
 			Current Deployment ID: Galaxy-Class
 
@@ -6843,7 +6843,7 @@ addEventListener('fetch', event => {});`
 			- Durable Objects:
 			  - EXAMPLE_DO_BINDING: ExampleDurableObject
 			Uploaded test-name (TIMINGS)
-			Deployed test-name triggers (TIMINGS)
+			Published test-name (TIMINGS)
 			  https://test-name.test-sub-domain.workers.dev
 			Current Deployment ID: Galaxy-Class
 
@@ -6887,7 +6887,7 @@ addEventListener('fetch', event => {});`
 			- Durable Objects:
 			  - EXAMPLE_DO_BINDING: ExampleDurableObject (defined in example-do-binding-worker)
 			Uploaded test-name (TIMINGS)
-			Deployed test-name triggers (TIMINGS)
+			Published test-name (TIMINGS)
 			  https://test-name.test-sub-domain.workers.dev
 			Current Deployment ID: Galaxy-Class
 
@@ -6936,7 +6936,7 @@ addEventListener('fetch', event => {});`
 			- Durable Objects:
 			  - EXAMPLE_DO_BINDING: ExampleDurableObject
 			Uploaded test-name (TIMINGS)
-			Deployed test-name triggers (TIMINGS)
+			Published test-name (TIMINGS)
 			  https://test-name.test-sub-domain.workers.dev
 			Current Deployment ID: Galaxy-Class
 
@@ -7061,7 +7061,7 @@ addEventListener('fetch', event => {});`
 			- Services:
 			  - FOO: foo-service - production
 			Uploaded test-name (TIMINGS)
-			Deployed test-name triggers (TIMINGS)
+			Published test-name (TIMINGS)
 			  https://test-name.test-sub-domain.workers.dev
 			Current Deployment ID: Galaxy-Class
 
@@ -7095,7 +7095,7 @@ addEventListener('fetch', event => {});`
 			- Analytics Engine Datasets:
 			  - FOO: foo-dataset
 			Uploaded test-name (TIMINGS)
-			Deployed test-name triggers (TIMINGS)
+			Published test-name (TIMINGS)
 			  https://test-name.test-sub-domain.workers.dev
 			Current Deployment ID: Galaxy-Class
 
@@ -7135,7 +7135,7 @@ addEventListener('fetch', event => {});`
 			- dispatch namespaces:
 			  - foo: Foo
 			Uploaded test-name (TIMINGS)
-			Deployed test-name triggers (TIMINGS)
+			Published test-name (TIMINGS)
 			  https://test-name.test-sub-domain.workers.dev
 			Current Deployment ID: Galaxy-Class
 
@@ -7196,7 +7196,7 @@ addEventListener('fetch', event => {});`
 			  - foo: Foo (outbound -> foo_outbound)
 			  - bar: Bar (outbound -> bar_outbound)
 			Uploaded test-name (TIMINGS)
-			Deployed test-name triggers (TIMINGS)
+			Published test-name (TIMINGS)
 			  https://test-name.test-sub-domain.workers.dev
 			Current Deployment ID: Galaxy-Class
 
@@ -7254,7 +7254,7 @@ addEventListener('fetch', event => {});`
 			- dispatch namespaces:
 			  - foo: Foo (outbound -> foo_outbound)
 			Uploaded test-name (TIMINGS)
-			Deployed test-name triggers (TIMINGS)
+			Published test-name (TIMINGS)
 			  https://test-name.test-sub-domain.workers.dev
 			Current Deployment ID: Galaxy-Class
 
@@ -7308,7 +7308,7 @@ addEventListener('fetch', event => {});`
 				  - something: \\"else\\"
 				  - nested: {\\"stuff\\":\\"here\\"}
 				Uploaded test-name (TIMINGS)
-				Deployed test-name triggers (TIMINGS)
+				Published test-name (TIMINGS)
 				  https://test-name.test-sub-domain.workers.dev
 				Current Deployment ID: Galaxy-Class
 
@@ -7349,7 +7349,7 @@ addEventListener('fetch', event => {});`
 				- Unsafe:
 				  - binding-type: my-binding
 				Uploaded test-name (TIMINGS)
-				Deployed test-name triggers (TIMINGS)
+				Published test-name (TIMINGS)
 				  https://test-name.test-sub-domain.workers.dev
 				Current Deployment ID: Galaxy-Class
 
@@ -7398,7 +7398,7 @@ addEventListener('fetch', event => {});`
 				- Unsafe:
 				  - plain_text: my-binding
 				Uploaded test-name (TIMINGS)
-				Deployed test-name triggers (TIMINGS)
+				Published test-name (TIMINGS)
 				  https://test-name.test-sub-domain.workers.dev
 				Current Deployment ID: Galaxy-Class
 
@@ -7440,7 +7440,7 @@ addEventListener('fetch', event => {});`
 					expect(std.out).toMatchInlineSnapshot(`
 				"Total Upload: xx KiB / gzip: xx KiB
 				Uploaded test-name (TIMINGS)
-				Deployed test-name triggers (TIMINGS)
+				Published test-name (TIMINGS)
 				  https://test-name.test-sub-domain.workers.dev
 				Current Deployment ID: Galaxy-Class
 
@@ -7548,7 +7548,7 @@ addEventListener('fetch', event => {});`
 					expect(std.out).toMatchInlineSnapshot(`
 				"Total Upload: xx KiB / gzip: xx KiB
 				Uploaded test-name (TIMINGS)
-				Deployed test-name triggers (TIMINGS)
+				Published test-name (TIMINGS)
 				  https://test-name.test-sub-domain.workers.dev
 				Current Deployment ID: Galaxy-Class
 
@@ -7594,7 +7594,7 @@ addEventListener('fetch', event => {});`
 			expect(std.out).toMatchInlineSnapshot(`
 			"Total Upload: xx KiB / gzip: xx KiB
 			Uploaded test-name (TIMINGS)
-			Deployed test-name triggers (TIMINGS)
+			Published test-name (TIMINGS)
 			  https://test-name.test-sub-domain.workers.dev
 			Current Deployment ID: Galaxy-Class
 
@@ -7627,7 +7627,7 @@ addEventListener('fetch', event => {});`
 			expect(std.out).toMatchInlineSnapshot(`
 			"Total Upload: xx KiB / gzip: xx KiB
 			Uploaded test-name (TIMINGS)
-			Deployed test-name triggers (TIMINGS)
+			Published test-name (TIMINGS)
 			  https://test-name.test-sub-domain.workers.dev
 			Current Deployment ID: Galaxy-Class
 
@@ -7664,7 +7664,7 @@ addEventListener('fetch', event => {});`
 			expect(std.out).toMatchInlineSnapshot(`
 			"Total Upload: xx KiB / gzip: xx KiB
 			Uploaded test-name (TIMINGS)
-			Deployed test-name triggers (TIMINGS)
+			Published test-name (TIMINGS)
 			  https://test-name.test-sub-domain.workers.dev
 			Current Deployment ID: Galaxy-Class
 
@@ -7717,7 +7717,7 @@ addEventListener('fetch', event => {});`
 			expect(std.out).toMatchInlineSnapshot(`
 			"Total Upload: xx KiB / gzip: xx KiB
 			Uploaded test-name (TIMINGS)
-			Deployed test-name triggers (TIMINGS)
+			Published test-name (TIMINGS)
 			  https://test-name.test-sub-domain.workers.dev
 			Current Deployment ID: Galaxy-Class
 
@@ -7819,7 +7819,7 @@ addEventListener('fetch', event => {});`
 			expect(std.out).toMatchInlineSnapshot(`
 			"Total Upload: xx KiB / gzip: xx KiB
 			Uploaded test-name (TIMINGS)
-			Deployed test-name triggers (TIMINGS)
+			Published test-name (TIMINGS)
 			  https://test-name.test-sub-domain.workers.dev
 			Current Deployment ID: Galaxy-Class
 
@@ -7852,7 +7852,7 @@ addEventListener('fetch', event => {});`
 			expect(std.out).toMatchInlineSnapshot(`
 			"Total Upload: xx KiB / gzip: xx KiB
 			Uploaded test-name (TIMINGS)
-			Deployed test-name triggers (TIMINGS)
+			Published test-name (TIMINGS)
 			  https://test-name.test-sub-domain.workers.dev
 			Current Deployment ID: Galaxy-Class
 
@@ -7890,7 +7890,7 @@ addEventListener('fetch', event => {});`
 				expect(std.out).toMatchInlineSnapshot(`
 			"Total Upload: xx KiB / gzip: xx KiB
 			Uploaded test-name (TIMINGS)
-			Deployed test-name triggers (TIMINGS)
+			Published test-name (TIMINGS)
 			  https://test-name.test-sub-domain.workers.dev
 			Current Deployment ID: Galaxy-Class
 
@@ -7924,7 +7924,7 @@ addEventListener('fetch', event => {});`
 			expect(std.out).toMatchInlineSnapshot(`
 			"Total Upload: xx KiB / gzip: xx KiB
 			Uploaded test-name (TIMINGS)
-			Deployed test-name triggers (TIMINGS)
+			Published test-name (TIMINGS)
 			  https://test-name.test-sub-domain.workers.dev
 			Current Deployment ID: Galaxy-Class
 
@@ -7957,7 +7957,7 @@ addEventListener('fetch', event => {});`
 			expect(std.out).toMatchInlineSnapshot(`
 			"Total Upload: xx KiB / gzip: xx KiB
 			Uploaded test-name (TIMINGS)
-			Deployed test-name triggers (TIMINGS)
+			Published test-name (TIMINGS)
 			  https://test-name.test-sub-domain.workers.dev
 			Current Deployment ID: Galaxy-Class
 
@@ -7992,7 +7992,7 @@ addEventListener('fetch', event => {});`
 			expect(std.out).toMatchInlineSnapshot(`
 			"Total Upload: xx KiB / gzip: xx KiB
 			Uploaded test-name (TIMINGS)
-			Deployed test-name triggers (TIMINGS)
+			Published test-name (TIMINGS)
 			  https://test-name.test-sub-domain.workers.dev
 			Current Deployment ID: Galaxy-Class
 
@@ -8025,7 +8025,7 @@ addEventListener('fetch', event => {});`
 			expect(std.out).toMatchInlineSnapshot(`
 			"Total Upload: xx KiB / gzip: xx KiB
 			Uploaded test-name (TIMINGS)
-			Deployed test-name triggers (TIMINGS)
+			Published test-name (TIMINGS)
 			  https://test-name.test-sub-domain.workers.dev
 			Current Deployment ID: Galaxy-Class
 
@@ -8071,7 +8071,7 @@ addEventListener('fetch', event => {});`
 			  "info": "",
 			  "out": "Total Upload: xx KiB / gzip: xx KiB
 			Uploaded test-name (TIMINGS)
-			Deployed test-name triggers (TIMINGS)
+			Published test-name (TIMINGS)
 			  https://test-name.test-sub-domain.workers.dev
 			Current Deployment ID: Galaxy-Class
 
@@ -8125,7 +8125,7 @@ addEventListener('fetch', event => {});`
 			  "info": "",
 			  "out": "Total Upload: xx KiB / gzip: xx KiB
 			Uploaded test-name (TIMINGS)
-			Deployed test-name triggers (TIMINGS)
+			Published test-name (TIMINGS)
 			  https://test-name.test-sub-domain.workers.dev
 			Current Deployment ID: Galaxy-Class
 
@@ -8153,7 +8153,7 @@ addEventListener('fetch', event => {});`
 			  "info": "",
 			  "out": "Total Upload: xx KiB / gzip: xx KiB
 			Uploaded test-name (TIMINGS)
-			Deployed test-name triggers (TIMINGS)
+			Published test-name (TIMINGS)
 			  https://test-name.test-sub-domain.workers.dev
 			Current Deployment ID: Galaxy-Class
 
@@ -8197,7 +8197,7 @@ addEventListener('fetch', event => {});`
 			  "out": "↗️  Done syncing assets
 			Total Upload: xx KiB / gzip: xx KiB
 			Uploaded test-name (TIMINGS)
-			Deployed test-name triggers (TIMINGS)
+			Published test-name (TIMINGS)
 			  https://test-name.test-sub-domain.workers.dev
 			Current Deployment ID: Galaxy-Class
 
@@ -8258,7 +8258,7 @@ export default{
 			  "info": "",
 			  "out": "Total Upload: xx KiB / gzip: xx KiB
 			Uploaded test-name (TIMINGS)
-			Deployed test-name triggers (TIMINGS)
+			Published test-name (TIMINGS)
 			  https://test-name.test-sub-domain.workers.dev
 			Current Deployment ID: Galaxy-Class
 
@@ -8495,7 +8495,7 @@ export default{
 			  "info": "",
 			  "out": "Total Upload: xx KiB / gzip: xx KiB
 			Uploaded test-name (TIMINGS)
-			Deployed test-name triggers (TIMINGS)
+			Published test-name (TIMINGS)
 			  https://test-name.test-sub-domain.workers.dev
 			Current Deployment ID: Galaxy-Class
 
@@ -9021,7 +9021,7 @@ export default{
 			- Queues:
 			  - QUEUE_ONE: queue1
 			Uploaded test-name (TIMINGS)
-			Deployed test-name triggers (TIMINGS)
+			Published test-name (TIMINGS)
 			  https://test-name.test-sub-domain.workers.dev
 			  Producer for queue1
 			Current Deployment ID: Galaxy-Class
@@ -9102,7 +9102,7 @@ export default{
 			expect(std.out).toMatchInlineSnapshot(`
 			"Total Upload: xx KiB / gzip: xx KiB
 			Uploaded test-name (TIMINGS)
-			Deployed test-name triggers (TIMINGS)
+			Published test-name (TIMINGS)
 			  https://test-name.test-sub-domain.workers.dev
 			  Consumer for queue1
 			Current Deployment ID: Galaxy-Class
@@ -9146,7 +9146,7 @@ export default{
 			expect(std.out).toMatchInlineSnapshot(`
 			"Total Upload: xx KiB / gzip: xx KiB
 			Uploaded test-name (TIMINGS)
-			Deployed test-name triggers (TIMINGS)
+			Published test-name (TIMINGS)
 			  https://test-name.test-sub-domain.workers.dev
 			  Consumer for queue1
 			Current Deployment ID: Galaxy-Class
@@ -9214,7 +9214,7 @@ export default{
 			expect(std.out).toMatchInlineSnapshot(`
 			"Total Upload: xx KiB / gzip: xx KiB
 			Uploaded test-name (TIMINGS)
-			Deployed test-name triggers (TIMINGS)
+			Published test-name (TIMINGS)
 			  https://test-name.test-sub-domain.workers.dev
 			  Consumer for queue1
 			Current Deployment ID: Galaxy-Class
@@ -9256,7 +9256,7 @@ export default{
 			expect(std.out).toMatchInlineSnapshot(`
 			"Total Upload: xx KiB / gzip: xx KiB
 			Uploaded test-name (TIMINGS)
-			Deployed test-name triggers (TIMINGS)
+			Published test-name (TIMINGS)
 			  https://test-name.test-sub-domain.workers.dev
 			  Consumer for queue1
 			Current Deployment ID: Galaxy-Class
@@ -9297,7 +9297,7 @@ export default{
 			expect(std.out).toMatchInlineSnapshot(`
 			"Total Upload: xx KiB / gzip: xx KiB
 			Uploaded test-name (TIMINGS)
-			Deployed test-name triggers (TIMINGS)
+			Published test-name (TIMINGS)
 			  https://test-name.test-sub-domain.workers.dev
 			  Consumer for queue1
 			Current Deployment ID: Galaxy-Class
@@ -9384,7 +9384,7 @@ export default{
 			- AI:
 			  - Name: AI_BIND
 			Uploaded test-name (TIMINGS)
-			Deployed test-name triggers (TIMINGS)
+			Published test-name (TIMINGS)
 			  https://test-name.test-sub-domain.workers.dev
 			Current Deployment ID: Galaxy-Class
 
@@ -9422,7 +9422,7 @@ export default{
 			└──────────────────────────────────────┴────────┴──────────┘
 			Total Upload: xx KiB / gzip: xx KiB
 			Uploaded test-name (TIMINGS)
-			Deployed test-name triggers (TIMINGS)
+			Published test-name (TIMINGS)
 			  https://test-name.test-sub-domain.workers.dev
 			Current Deployment ID: Galaxy-Class
 
@@ -9456,7 +9456,7 @@ export default{
 			└──────────────────────────────────────┴────────┴──────────┘
 			Total Upload: xx KiB / gzip: xx KiB
 			Uploaded test-name (TIMINGS)
-			Deployed test-name triggers (TIMINGS)
+			Published test-name (TIMINGS)
 			  https://test-name.test-sub-domain.workers.dev
 			Current Deployment ID: Galaxy-Class
 
@@ -9495,7 +9495,7 @@ export default{
 			- Hyperdrive Configs:
 			  - HYPERDRIVE: 343cd4f1d58c42fbb5bd082592fd7143
 			Uploaded test-name (TIMINGS)
-			Deployed test-name triggers (TIMINGS)
+			Published test-name (TIMINGS)
 			  https://test-name.test-sub-domain.workers.dev
 			Current Deployment ID: Galaxy-Class
 
@@ -9529,7 +9529,7 @@ export default{
 			- mTLS Certificates:
 			  - CERT_ONE: 1234
 			Uploaded test-name (TIMINGS)
-			Deployed test-name triggers (TIMINGS)
+			Published test-name (TIMINGS)
 			  https://test-name.test-sub-domain.workers.dev
 			Current Deployment ID: Galaxy-Class
 
@@ -9558,7 +9558,7 @@ export default{
 			expect(std.out).toMatchInlineSnapshot(`
 			"Total Upload: xx KiB / gzip: xx KiB
 			Uploaded test-name (TIMINGS)
-			Deployed test-name triggers (TIMINGS)
+			Published test-name (TIMINGS)
 			  https://test-name.test-sub-domain.workers.dev
 			Current Deployment ID: Galaxy-Class
 
@@ -9586,7 +9586,7 @@ export default{
 			expect(std.out).toMatchInlineSnapshot(`
 			"Total Upload: xx KiB / gzip: xx KiB
 			Uploaded test-name (TIMINGS)
-			Deployed test-name triggers (TIMINGS)
+			Published test-name (TIMINGS)
 			  https://test-name.test-sub-domain.workers.dev
 			Current Deployment ID: Galaxy-Class
 
@@ -9616,7 +9616,7 @@ export default{
 			expect(std.out).toMatchInlineSnapshot(`
 			"Total Upload: xx KiB / gzip: xx KiB
 			Uploaded test-name (TIMINGS)
-			Deployed test-name triggers (TIMINGS)
+			Published test-name (TIMINGS)
 			  https://test-name.test-sub-domain.workers.dev
 			Current Deployment ID: Galaxy-Class
