+++ packages/wrangler/src/__tests__/versions/versions.help.test.ts
@@ -60,8 +60,6 @@ describe("versions --help", () => {
 		expect(std.out).toMatchInlineSnapshot(`
 		"wrangler versions
 
-		List, view, upload and deploy Versions of your Worker to Cloudflare [beta]
-
 		Commands:
 		  wrangler versions view <version-id>         View the details of a specific version of your Worker [beta]
 		  wrangler versions list                      List the 10 most recent Versions of your Worker [beta]
@@ -98,8 +96,6 @@ describe("versions subhelp", () => {
 		expect(std.out).toMatchInlineSnapshot(`
 		"wrangler versions
 
-		List, view, upload and deploy Versions of your Worker to Cloudflare [beta]
-
 		Commands:
 		  wrangler versions view <version-id>         View the details of a specific version of your Worker [beta]
 		  wrangler versions list                      List the 10 most recent Versions of your Worker [beta]
@@ -124,8 +120,6 @@ describe("versions subhelp", () => {
 		expect(std.out).toMatchInlineSnapshot(`
 		"wrangler versions
 
-		List, view, upload and deploy Versions of your Worker to Cloudflare [beta]
-
 		Commands:
 		  wrangler versions view <version-id>         View the details of a specific version of your Worker [beta]
 		  wrangler versions list                      List the 10 most recent Versions of your Worker [beta]
@@ -149,8 +143,6 @@ describe("versions subhelp", () => {
 
 		expect(std.out).toMatchInlineSnapshot(`
 		"wrangler versions
-
-		List, view, upload and deploy Versions of your Worker to Cloudflare [beta]
 
 		Commands:
 		  wrangler versions view <version-id>         View the details of a specific version of your Worker [beta]
