+++ packages/wrangler/src/api/dev.ts
@@ -211,7 +211,6 @@ export async function unstable_dev(
 		logLevel: options?.logLevel ?? defaultLogLevel,
 		port: options?.port ?? 0,
 		updateCheck: options?.updateCheck ?? false,
-		experimentalVersions: undefined,
 	};
 
 	//due to Pages adoption of unstable_dev, we can't *just* disable rebuilds and watching. instead, we'll have two versions of startDev, which will converge.
