+++ packages/wrangler/src/generate/index.ts
@@ -55,7 +55,6 @@ export async function generateHandler(args: GenerateArgs) {
 			_: args._,
 			$0: args.$0,
 			experimentalJsonConfig: false,
-			experimentalVersions: args.experimentalVersions,
 		});
 	}
