+++ packages/wrangler/src/metrics/send-event.ts
@@ -72,8 +72,7 @@ export type EventNames =
 	| "view versioned deployment"
 	| "view latest versioned deployment"
 	| "list versioned deployments"
-	| "download pages config"
-	| "deploy worker triggers";
+	| "download pages config";
 
 /**
  * Send a metrics event, with no extra properties, to Cloudflare, if usage tracking is enabled.
