+++ packages/wrangler/src/yargs-types.ts
@@ -9,7 +9,6 @@ export interface CommonYargsOptions {
 	config: string | undefined;
 	env: string | undefined;
 	"experimental-json-config": boolean | undefined;
-	"experimental-versions": boolean | undefined;
 }
 
 /**
